Rename Mongo impls; extract neutral MongoRepositoryBase (#2877)
Mirror the impl names to the interfaces, and avoid the contradictory
"UserRestricted : Unrestricted" inheritance the literal rename would produce:

- MongoRepositoryBase (abstract): the shared user/journal/entry data access,
  with GetAllJournalDocumentsFilter as an abstract read-shaping hook.
- UnrestrictedMongoRepository : MongoRepositoryBase, IUnrestrictedRepository -
  no filter + the maintenance ops (WakeMeUp, CountAll*). (was MongoRepository)
- UserRestrictedMongoRepository : MongoRepositoryBase, IUserRestrictedRepository
  - the per-user filter + write guards + CurrentUser. (was UserScopedMongoRepository)

The two seams are now siblings (neither extends the other), matching the
symmetric interface model. Because IUserRestrictedRepository excludes
IMaintenanceRepository, the scoped repo no longer even has WakeMeUp/CountAll;
a few tests that pinged WakeMeUp on a scoped repo as a redundant warm-up had
those no-op calls removed.

Behavior-preserving. Full solution builds clean; Core 76, EphemeralMongo 95 and
Api 17 (incl. the DI smoke tests, which confirm the new types resolve) green.
