+++ api/Engraved.Persistence.Mongo.Tests/Source/MongoRepository_GetAllJournals_Should.cs
@@ -12,7 +12,7 @@ namespace Engraved.Persistence.Mongo.Tests;
 public class MongoRepository_GetAllJournals_Should
 {
   private string _gaugeJournalId = null!;
-  private MongoRepository _repository = null!;
+  private UnrestrictedMongoRepository _repository = null!;
 
   [SetUp]
   public async Task Setup()
