+++ api/Engraved.Persistence.Mongo.Tests/Source/MongoRepository_SearchEntries_Should.cs
@@ -13,7 +13,7 @@ namespace Engraved.Persistence.Mongo.Tests;
 public class MongoRepository_SearchEntries_Should
 {
   private string _journalId = null!;
-  private MongoRepository _repository = null!;
+  private UnrestrictedMongoRepository _repository = null!;
 
   [SetUp]
   public async Task Setup()
