+++ api/Engraved.Persistence.Mongo.Tests/Source/UserScopedMongoRepository_Permissions_Should.cs
@@ -17,8 +17,8 @@ public class UserScopedMongoRepository_Permissions_Should
 
   private const string OtherUserName = "other";
   private string _otherUserId = null!;
-  private MongoRepository _repository = null!;
-  private UserScopedMongoRepository _userScopedRepository = null!;
+  private UnrestrictedMongoRepository _repository = null!;
+  private UserRestrictedMongoRepository _userScopedRepository = null!;
 
   [SetUp]
   public async Task Setup()
